Add getRecyclingZoneById

+++ src/main/java/com/example/proyectoFinal1/dto/RecyclingZoneDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class RecyclingZoneDTO {
-    private int recyclingZoneId;
+    private Long recyclingZoneId;
     private String RecyclingZoneName;
     private String RecyclingZoneAdress;
